Доделываю Service-ы и controller-ы исправление ошибок

+++ src/main/java/ru/alex/two/repository/TypesRepository.java
@@ -3,5 +3,5 @@ package ru.alex.two.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.alex.two.domain.Types;
 
-public interface TypesRepository  extends JpaRepository<Types, Integer> {
+public interface TypesRepository  extends JpaRepository<Types, Long> {
 }
